refactor: move sidePanel and SearchBar into containers folder
